% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Create data directory}
\usage{
getdata(dir = "data", dataset = c("all", "dataskills"))
}
\arguments{
\item{dir}{The name of the directory to save the data in (default "data")}

\item{dataset}{Which dataset to download, from "all", "dataskills" (default "all")}
}
\value{
A list of the paths of the new data files
}
\description{
Save data files to a directory
}
